fix: fix cyclic deps (#588)
* fix: fix cyclic deps

* refactor: update store exports

* chore(package): update deps

* refactor: refactor shared module

* fix: fix remaining cyclic deps
