+++ package.json
@@ -44,15 +44,15 @@
     "express": "~4.16.4",
     "debug": "~4.1.0",
     "compression": "~1.7.3",
-    "@angular/animations": "~7.1.0",
-    "@angular/common": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
-    "@angular/http": "~7.1.0",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/platform-server": "~7.1.0",
-    "@angular/router": "~7.1.0",
+    "@angular/animations": "~7.1.1",
+    "@angular/common": "~7.1.1",
+    "@angular/core": "~7.1.1",
+    "@angular/forms": "~7.1.1",
+    "@angular/http": "~7.1.1",
+    "@angular/platform-browser": "~7.1.1",
+    "@angular/platform-browser-dynamic": "~7.1.1",
+    "@angular/platform-server": "~7.1.1",
+    "@angular/router": "~7.1.1",
     "@angular/cdk": "~7.1.0",
     "@angular/material": "~7.1.0",
     "@angular/flex-layout": "7.0.0-beta.19",
@@ -81,14 +81,14 @@
     "@ngx-translate/http-loader": "~4.0.0",
     "@ngx-meta/core": "~6.0.0",
     "unionize": "~2.1.2",
-    "angulartics2": "~7.2.0",
+    "angulartics2": "~7.2.1",
     "ngx-perfect-scrollbar": "~7.1.0"
   },
   "devDependencies": {
     "@angular/cli": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/compiler-cli": "~7.1.0",
-    "@angular/language-service": "~7.1.0",
+    "@angular/compiler": "~7.1.1",
+    "@angular/compiler-cli": "~7.1.1",
+    "@angular/language-service": "~7.1.1",
     "@angular-devkit/build-angular": "~0.11.0",
     "@angular-devkit/core": "~7.1.0",
     "@angular-devkit/schematics": "~7.1.0",
@@ -99,7 +99,7 @@
     "@ngrx/store-devtools": "~6.1.2",
     "@types/compression": "~0.0.36",
     "@types/express": "~4.16.0",
-    "@types/jest": "~23.3.9",
+    "@types/jest": "~23.3.10",
     "@types/lodash": "~4.14.118",
     "@types/nightmare": "~2.10.4",
     "@types/node": "~10.12.10",
