+++ src/app/framework/core/base-container.component.ts
@@ -5,9 +5,11 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { Observable } from 'rxjs';
 import { Store } from '@ngrx/store';
 
+// app
+import { FrameworkState } from '~/app/framework/store';
+
 // module
 import { BaseComponent } from './base.component';
-import { State } from '~/app/store';
 
 @Component({
   template: '',
@@ -17,7 +19,7 @@ export class BaseContainerComponent extends BaseComponent {
   error$: Observable<string>;
   isProcessing$: Observable<boolean>;
 
-  constructor(protected readonly store$: Store<State>) {
+  constructor(protected readonly store$: Store<FrameworkState>) {
     super();
   }
 }
