+++ src/app/framework/i18n/index.ts
@@ -1,4 +1,2 @@
-export { initialLanguage, Language } from './models/language';
-export { ChangeLanguageComponent } from './change-language.component';
-export { I18NService } from './i18n.service';
-export { I18NModule, translateFactory } from './i18n.module';
+export * from './change-language.component';
+export * from './i18n.module';
