+++ src/app/framework/i18n/testing/index.ts
@@ -1,1 +1,1 @@
-export { I18NTestingModule } from './i18n-testing.module';
+export * from './i18n-testing.module';
