+++ src/app/framework/material/index.ts
@@ -1,1 +1,1 @@
-export { MaterialModule } from './material.module';
+export * from './material.module';
