+++ src/app/framework/ngrx/constants.ts
@@ -1,0 +1,2 @@
+// HTTP client
+export const HTTP_CLIENT__MAX_RETRIES = 3;
