+++ src/app/framework/ngrx/testing/index.ts
@@ -1,3 +1,4 @@
-export { MockEntityService } from './mocks/entity-service.mock';
-export { NgrxTestingModule } from './ngrx-testing.module';
-export { getState } from './state.util';
+export * from './mocks/entity-service.mock';
+export * from './mocks/resource.mock';
+export * from './ngrx-testing.module';
+export * from './state.util';
