+++ src/app/framework/ngrx/testing/mocks/resource.mock.ts
@@ -1,0 +1,1 @@
+export type MockResource = any;
