+++ src/app/framework/store/i18n/index.ts
@@ -2,8 +2,8 @@
 import * as LanguageSelectors from './language/language.selectors';
 
 export { languageActions } from './language/language.actions';
-export { Language } from './language/language.model';
+export * from './language/language.model';
 export { LanguageSelectors };
 
-export { I18NModule } from './i18n.module';
-export { State } from './i18n.state';
+export { I18NModule as I18NStoreModule } from './i18n.module';
+export { State as I18NState } from './i18n.state';
