+++ src/app/framework/store/i18n/language/testing/common.ts
@@ -1,4 +1,4 @@
 // app
-import { initialLanguage } from '~/app/framework/i18n';
+import { initialLanguage } from '~/app/framework/store';
 
 export const MOCK_LANGUAGE = initialLanguage;
