+++ src/app/framework/store/index.ts
@@ -1,3 +1,3 @@
 export * from './i18n';
-export { State } from './state';
+export { State as FrameworkState } from './state';
 export { StoreModule as StoreFrameworkModule } from './store.module';
