+++ src/app/framework/store/state.ts
@@ -1,5 +1,5 @@
 // module
-import { State as I18NState } from './i18n';
+import { I18NState } from './i18n';
 
 export interface State {
   i18n: I18NState;
