+++ src/app/framework/testing/index.ts
@@ -3,5 +3,5 @@ import { ngJest } from './shorthand/ng-jest';
 
 export const t = ngJest;
 
-export { MockComponent } from './mocks/component.mock';
-export { TestingModule } from './testing.module';
+export * from './mocks/component.mock';
+export * from './testing.module';
