+++ src/app/shared/constants.ts
@@ -1,6 +1,3 @@
-// Http client
-export const HTTP_CLIENT__MAX_RETRIES = 3;
-
 // Application errors
 export const ERROR__NO_PAYLOAD = {
   status: 0,
