+++ src/app/store/air-universal/airline/testing/index.ts
@@ -1,2 +1,2 @@
-export { MockAirlineService } from './mocks/airline-service.mock';
-export { MOCK_AIRLINE, MOCK_AIRLINES } from './common';
+export * from './mocks/airline-service.mock';
+export * from './common';
