+++ src/app/store/air-universal/index.ts
@@ -2,9 +2,9 @@
 import * as AirlineSelectors from './airline/airline.selectors';
 
 export { airlineActions } from './airline/airline.actions';
-export { Airline, initialAirline } from './airline/airline.model';
+export * from './airline/airline.model';
 export { AIRLINE } from './airline/airline.state';
 export { AirlineSelectors };
 
-export { AirUniversalModule } from './air-universal.module';
+export { AirUniversalModule as AirUniversalStoreModule } from './air-universal.module';
 export { State as AirUniversalState } from './air-universal.state';
