+++ src/app/store/index.ts
@@ -1,4 +1,4 @@
 export * from '~/app/framework/store';
 export * from './air-universal';
-export { State } from './state';
-export { StoreModule } from './store.module';
+export * from './state';
+export * from './store.module';
