+++ src/app/store/state.ts
@@ -1,6 +1,9 @@
+// app
+import { FrameworkState } from '~/app/framework/store';
+
 // module
 import { AirUniversalState } from './air-universal';
 
-export interface State {
+export interface State extends FrameworkState {
   airUniversal: AirUniversalState;
 }
